+++ rust/src/structure_designer/common_constants.rs
@@ -4,4 +4,4 @@ use glam::i32::IVec3;
 pub const IMPLICIT_VOLUME_MIN: IVec3 = IVec3::new(-10, -10, -10);
 pub const IMPLICIT_VOLUME_MAX: IVec3 = IVec3::new(10, 10, 10);
 
-pub const DIAMOND_UNIT_CELL_SIZE_ANGSTROM: f32 = 3.567;  // Size of one complete unit cell in Ångströms
+pub const DIAMOND_UNIT_CELL_SIZE_ANGSTROM: f64 = 3.567;  // Size of one complete unit cell in Ångströms
